capstone-website/requirements.txt
streamlit
Pillow
altair
streamlit_lottie
requests
numpy
pandas
python-time
rdkit-pypi
streamlit_extras
